+++ roles/bitrise-den/tests/test_bitrise_cli.py
@@ -1,5 +1,5 @@
 def test_bitrise_den_agent(host):
-    bitrise_cli = host.file("/Users/vagrant/bitrise-den-agent")
-    assert bitrise_cli.user == "linuxbrew"
-    assert bitrise_cli.group == "staff"
-    assert bitrise_cli.mode == 0o755
+    bitrise_den_agent = host.file("/home/linuxbrew/bitrise-den-agent")
+    assert bitrise_den_agent.user == "linuxbrew"
+    assert bitrise_den_agent.group == "staff"
+    assert bitrise_den_agent.mode == 0o755